source/applications/lib/text-embedding/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
    # via
    #   text-embedding (pyproject.toml)
    #   domain
    # via text-embedding (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
charset-normalizer==3.4.4
    # via core
deprecated==1.3.1
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-semantic-conventions
googleapis-common-protos==1.72.0
    # via opentelemetry-exporter-otlp-proto-grpc
grpcio==1.75.1
    # via
    #   text-embedding (pyproject.toml)
    #   grpcio-tools
    #   opentelemetry-exporter-otlp-proto-grpc
grpcio-tools==1.71.2
    # via text-embedding (pyproject.toml)
importlib-metadata==8.6.1
    # via opentelemetry-api
opentelemetry-api==1.32.1
    # via
    #   core
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.32.1
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.32.1
    # via core
opentelemetry-instrumentation==0.53b1
    # via opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-httpx==0.53b1
    # via core
opentelemetry-proto==1.32.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.32.1
    # via
    #   core
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.53b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-sdk
opentelemetry-util-http==0.53b1
    # via opentelemetry-instrumentation-httpx
packaging==25.0
    # via opentelemetry-instrumentation
protobuf==5.29.5
    # via
    #   text-embedding (pyproject.toml)
    #   googleapis-common-protos
    #   grpcio-tools
    #   opentelemetry-proto
pydantic==2.11.10
    # via
    #   text-embedding (pyproject.toml)
    #   core
    #   domain
pydantic-core==2.33.2
    # via pydantic
setuptools==80.9.0
    # via grpcio-tools
typing-extensions==4.15.0
    # via
    #   grpcio
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
wrapt==1.17.3
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
zipp==3.23.0
    # via importlib-metadata
